Fix positioning in Chrome at fractional zoom levels
Summary:
Chrome/WebKit's zooming can cause rounding errors around 0.01px.
The svgs had a width/height of 1px and an implicit "0 0 1 1" viewbox,
with the annotations rendered in overflow. This setup means that errors
of 0.01px really matter, since the rounding messes with the scale
of the whole SVG. This change makes annotations be 100x100, where
rounding errors of 0.01px don't matter.

Before:
{F1410866}

After:
{F1410865}

Reviewers: joshua, russell

Reviewed By: russell

Differential Revision: https://phabricator.khanacademy.org/D53305